LPD-61930 Rename and move to the correct package
